PR2: extract EvolutionTimeline into pure component + thin authed container
